+++ src/lintsource.ts
@@ -15,6 +15,7 @@ export type LintSource = ((state: EditorState) => Diagnostic[] | Promise<Diagnos
 	// eslint-disable-next-line @typescript-eslint/method-signature-style
 	fixer?: (doc: Text, rule?: string) => string | Promise<string>;
 };
+export type LintSources = LintSource | [LintSource, LintSource?];
 export type LintSourceGetter = (
 	opt?: Option | LiveOption,
 	view?: EditorView,